rag-eval-pipeline/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-milvus==0.1.4
openai==1.51.0

langfuse==2.53.0

fastmcp==0.1.0

chromadb==0.5.15

# Optional: swap chroma for milvus
# pymilvus==2.4.6
